+++ ClickHaskell/DeSerialization.hs
@@ -1,8 +1,0 @@
-
-module ClickHaskell.DeSerialization
-  {-# DEPRECATED "\
-  \This module would be deleted in next minor release\
-  \Please don't use it's imports\
-  \" #-}
-  () where
-