+++ ChangeLog.md
@@ -9,10 +9,39 @@ ClickHaskell documentation got it's own domain name: https://clickhaskell.dev/
 ## Features:
 - Query serialization support for UUID (+ add test)
 - Export of client errors for exception handling
+- Dropped vector dependency
+- Introduced memory consumption test (64M limit) on parallel reading and writing of 1 million rows
 
 ## API changes:
+- ### Migration to streaming API
+    The result of selects now exposes the block by block handling result. So you need to pass the handler and to process the list of results
+    ```haskell
+    result <-
+      sum <$>
+        select
+          @ExampleColumns
+          @ExampleData
+          connection "\
+          \SELECT * \
+          \FROM generateRandom('a1 Int64', 1, 10, 2) \
+          \LIMIT 1_000_00 \
+          \ "
+    ```
+    now looks like
+    ```haskell
+    result <-
+      sum <$>
+        select
+          @ExampleColumns
+          @ExampleData
+          connection "\
+          \SELECT * \
+          \FROM generateRandom('a1 Int64', 1, 10, 2) \
+          \LIMIT 1_000_00 \
+          \ "
+          (pure . sum)
+    ```
 - ### DateTime type now parametrized with timezone
-    Migration guide:\
     Every DateTime type annotations
     ```haskell
     type A = ChDateTime -- DateTime
@@ -25,7 +54,6 @@ ClickHaskell documentation got it's own domain name: https://clickhaskell.dev/
     ```
 
 - ### Migration to single module distribution
-    Migration guide: \
     You need to move all imports such as
     ```haskell
     import ClickHaskell.DbTypes (ChInt8)